requirements.txt
fastapi==0.95.1
pydantic==1.10.7
python-dotenv==1.0.0
mongoengine==0.27.0
bcrypt==4.0.1
cryptography==41.0.3
PyJWT==2.8.0
uvicorn==0.22.0
